+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/components/SelectAssetFlow.tsx
@@ -7,7 +7,6 @@ import Fuse from "fuse.js";
 import { useCurrenciesByMarketcap } from "@ledgerhq/live-common/currencies/hooks";
 import Text from "~/renderer/components/Text";
 import { SelectNetwork } from "./SelectNetwork";
-import { listAndFilterCurrencies } from "@ledgerhq/live-common/platform/helpers";
 import { getEnv } from "@ledgerhq/live-env";
 
 const options = {
@@ -23,21 +22,14 @@ function fuzzySearch(currencies: Currency[], searchValue: string): Currency[] {
 }
 
 export type SelectAccountAndCurrencyDrawerProps = {
-  assetIds?: string[];
-  includeTokens?: boolean;
   onAssetSelected: (asset: CryptoOrTokenCurrency) => void;
+  currencies: CryptoOrTokenCurrency[];
 };
 
 function SelectAssetFlow(props: SelectAccountAndCurrencyDrawerProps) {
-  const { onAssetSelected, assetIds, includeTokens } = props;
-
+  const { onAssetSelected, currencies } = props;
   const { t } = useTranslation();
   const [searchValue, setSearchValue] = useState<string>("");
-
-  const currencies = listAndFilterCurrencies({
-    currencies: assetIds,
-    includeTokens,
-  });
 
   // sorting them by marketcap
   const sortedCurrencies = useCurrenciesByMarketcap(currencies);
@@ -57,6 +49,13 @@ function SelectAssetFlow(props: SelectAccountAndCurrencyDrawerProps) {
     },
     [onAssetSelected],
   );
+
+  // TODO : this is a temporary to handle the case where we have only one asset and one currency
+  if (currencies.length === 1) {
+    const currency = currencies[0];
+    onAssetSelected(currency);
+    return null;
+  }
 
   return (
     <SelectAccountAndCurrencyDrawerContainer>