+++ libs/coin-modules/coin-tron/src/logic/listOperations.unit.test.ts
@@ -1,5 +1,5 @@
 import BigNumber from "bignumber.js";
-import { fetchTronAccountTxs } from "../network";
+import { defaultFetchParams, fetchTronAccountTxs, getBlock } from "../network";
 import { fromTrongridTxInfoToOperation } from "../network/trongrid/trongrid-adapters";
 import { TrongridTxInfo, TronAsset } from "../types";
 import { listOperations } from "./listOperations";
@@ -8,6 +8,7 @@ import type { Operation } from "@ledgerhq/coin-framework/api/index";
 // Mock the fetchTronAccountTxs and fromTrongridTxInfoToOperation functions
 jest.mock("../network", () => ({
   fetchTronAccountTxs: jest.fn(),
+  getBlock: jest.fn(),
 }));
 
 jest.mock("../network/trongrid/trongrid-adapters", () => ({
@@ -17,13 +18,27 @@ jest.mock("../network/trongrid/trongrid-adapters", () => ({
 describe("listOperations", () => {
   const mockAddress = "tronExampleAddress";
 
+  const mockBlockTime = new Date("2023-01-01T00:00:00Z");
+  const mockBlockTimestamp = mockBlockTime.getTime();
+  beforeEach(() => {
+    jest.clearAllMocks();
+    (getBlock as jest.Mock).mockResolvedValue({
+      time: mockBlockTime,
+    });
+  });
+
+  const expectedFetchParams = {
+    ...defaultFetchParams,
+    minTimestamp: mockBlockTimestamp,
+  };
+
   it("should fetch transactions and return operations", async () => {
     const mockTxs: Partial<TrongridTxInfo>[] = [
       { txID: "tx1", value: new BigNumber(0) },
       { txID: "tx2", value: new BigNumber(42) },
     ];
 
-    const mockOperations: Partial<Operation<TronAsset>>[] = [
+    const expectedOperations: Partial<Operation<TronAsset>>[] = [
       { tx: { hash: "tx1" } as Partial<Operation<TronAsset>>["tx"], value: BigInt(0) },
       { tx: { hash: "tx2" } as Partial<Operation<TronAsset>>["tx"], value: BigInt(42) },
     ];
@@ -36,25 +51,35 @@ describe("listOperations", () => {
       };
     });
 
-    const [operations, token] = await listOperations(mockAddress);
+    const [operations, token] = await listOperations(mockAddress, 0);
 
-    expect(fetchTronAccountTxs).toHaveBeenCalledWith(mockAddress, expect.any(Function), {});
+    expect(fetchTronAccountTxs).toHaveBeenCalledWith(
+      mockAddress,
+      expect.any(Function),
+      {},
+      expectedFetchParams,
+    );
     expect(fromTrongridTxInfoToOperation).toHaveBeenCalledTimes(mockTxs.length);
-    expect(operations).toEqual(mockOperations);
+    expect(operations).toEqual(expectedOperations);
     expect(token).toBe("");
   });
 
   it("should handle empty transactions", async () => {
     const mockTxs: Partial<TrongridTxInfo>[] = [];
-    const mockOperations: Partial<Operation<TronAsset>>[] = [];
+    const expectedOperations: Partial<Operation<TronAsset>>[] = [];
 
     (fetchTronAccountTxs as jest.Mock).mockResolvedValue(mockTxs);
     (fromTrongridTxInfoToOperation as jest.Mock).mockImplementation(() => null);
 
-    const [operations, token] = await listOperations(mockAddress);
+    const [operations, token] = await listOperations(mockAddress, 0);
 
-    expect(fetchTronAccountTxs).toHaveBeenCalledWith(mockAddress, expect.any(Function), {});
-    expect(operations).toEqual(mockOperations);
+    expect(fetchTronAccountTxs).toHaveBeenCalledWith(
+      mockAddress,
+      expect.any(Function),
+      {},
+      expectedFetchParams,
+    );
+    expect(operations).toEqual(expectedOperations);
     expect(token).toBe("");
   });
 
@@ -62,6 +87,6 @@ describe("listOperations", () => {
     const exampleError = new Error("Network error!");
     (fetchTronAccountTxs as jest.Mock).mockRejectedValue(exampleError);
 
-    await expect(listOperations(mockAddress)).rejects.toThrow(new Error(exampleError.message));
+    await expect(listOperations(mockAddress, 0)).rejects.toThrow(new Error(exampleError.message));
   });
 });