+++ libs/coin-modules/coin-vechain/jest.config.js
@@ -20,4 +20,5 @@ module.exports = {
       { outputName: "vechain-sonar-executionTests-report.xml", reportedFilePath: "absolute" },
     ],
   ],
+  setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };